requirements.txt
Pillow==8.2.0
pygame==2.0.1
pygame-gui==0.5.7
pygame-widgets==0.4.0
